add autotracker support
This commit adds autotracking to Hamsda's ZOOTR map and item tracker.
This work is based on codemann8's LTTP autotracking.lua script and takes
advantage of the Bizhawk connector.lua that is included with Emotracker.

This branch is tested and known to work great with the following:
* EmoTracker 2.3.8.14
* Hamsda commit 9b3285e
* BizHawk 2.3.0 (x64) commit af9d1db
* The bizhawk connector.lua included with aforementioned EmoTracker

Also works simultaneously with Archipelago! (ootMulti.lua 2021-09-11)

To enable autotracking, right-click the robot icon in the lower right
corner and select "SNES", then "Lua". You will also need to find the
connector.lua included with Emotracker and run it in Bizhawk.

It's always best to reset the tracker and ensure all of your settings
are configured correctly prior to activating the autotracker. It can
also be reactivated after you load your save file in EmoTracker.

Things that are currently autotracked:
* Inventory and quest items received
* Locations checked
* Dungeon bosses defeated
* Trade sequences progressed
* Keys received (for keysanity variants)

Things that are NOT autotracked:
* Master Quest dungeon checks
* Which prize is awarded by each dungeon
* Scarecrow's song state
* Captures for spawn and entrance randomizer
* At which song location each song was received

I couldn't have done this without this amazing list of offsets:
* https://github.com/RiptideSage/OoT-CompletedChecks

Thank you RiptideSage!

I also heavily referenced the following resources:
* https://docs.google.com
  /spreadsheets/d/1lh8JcUcuXg7GR4DTWMhmfk42Y0H6nJvUg-zRn8-8bPk
* https://tcrf.net
  /Proto:The_Legend_of_Zelda:_Ocarina_of_Time_Master_Quest/Event_Editor
* https://wiki.cloudmodding.com/oot/Save_Format
* https://wiki.cloudmodding.com/oot/Scene_Table/NTSC_1.0
* https://github.com/aliblong/zootr-autotracker
* https://cloudmodding.com/zelda/oot
* https://wiki.ootrandomizer.com/index.php?title=Editing_RAM

Things that should be reviewed prior to merging:

To enable autotracking within EmoTracker for this pack, the platform was
changed to "snes" in the manifest. I couldn't get the option to show up
without making this change, and I don't know if it has any side effects.

A proof-of-concept "race mode" toggle was added to demonstrate how users
might be encouraged to stay honest and obey community racing rules.
Autotracking is disabled while race mode is enabled.

A proof-of-concept big poe counter was also added to the settings to
allow the user to indicate how many are required. This was necessary in
order to automate the big poe turn-in check.

I renamed the 'Deku Scrubs / Sticks' location to remove the forward
slash because it was preventing lookup.

I also changed HC Malon's visibility rules to make her disappear
when the egg is placed into the inventory. This solved some issues
relating to the advancement of the trade sequence.

Wishlist for things I don't feel like working on:
* fix many location codes for entrance rando variants
* test composer grave song check when Song Locations is used
* figure out how to add MQ dungeon checks and status
* change the interface in an obvious way when not in race mode
* don't allow race mode to be restored after it was turned off
* track dungeon prizes when that information is known to the player
* track scarecrow song without giving away info player doesn't have yet
* rescue epona (no location for this on the tracker?)
* stop reading a scene if all checks in that scene are completed