expand accessibility for deku theater skull mask
The Deku Theater Skull Mask check will now appear accessible even after
you traded to a different mask, since you can go back and get the right
one. This change means that the item in this slot should be indicative
of the furthest trade progress you've made, not the item you're actually
holding, which is how the autotracker handles this slot.