add unit tests for autotracker
This commit adds tests for the unit testing framework "busted" to the
project under the "tests" folder. The intent here is to add unit tests
to cover the more troublesome elements of the autotracker, notably the
child trade sequence handler, but more tests can be added to cover any
part of the project.

The following command can be used to execute all tests from the top
directory, next to the .busted file:

  $ busted .

The .busted file is used to define the parameters that should be used
when the tests are run this way.