autotracker: fix child trade sequence advancement
This commit fixes an issue where the advancement of the trade sequence
would not take place when the guard is shown the letter while kak is
closed. This is because updates to the trade sequence took place only
during the inventory callback, which does not trigger on this event
since the inventory does not change.

The handling of the child trade sequence was refactored to allow it to
be triggered from the events callback, where an update of the guard's
state was added. This ensures the trade sequence is progressed
immediately after the guard is shown the letter, similarly to the way
Talon's wake up check is handled.

This commit also adds two unit tests to cover these two cases.