don't allow race mode to be turned on
There's no need for this toggle since community race rules don't
allow the use of trackers that have autotracking functionality,
even if the autotracker is turned off.

This commit removes the option to turn on race mode, in order to avoid
misleading the user into thinking it allows them to use the tracker
for races. The casual mode indicator is retained to help indicate that
this tracker is capable of autotracking.